Move memex.events to h.events
This is part of a project to reintegrate the `memex` package into the
main `h` package: https://notes.wtk.io/2017/03/08/reintegrating-memex.
