+++ h/indexer/__init__.py
@@ -9,4 +9,4 @@ __all__ = (
 
 def includeme(config):
     config.add_subscriber('h.indexer.subscribers.subscribe_annotation_event',
-                          'memex.events.AnnotationEvent')
+                          'h.events.AnnotationEvent')
