+++ h/nipsa/__init__.py
@@ -5,7 +5,7 @@ def includeme(config):
     # Register the transform_annotation subscriber so that nipsa fields are
     # written into annotations on save.
     config.add_subscriber('h.nipsa.subscribers.transform_annotation',
-                          'memex.events.AnnotationTransformEvent')
+                          'h.events.AnnotationTransformEvent')
 
     # Register an additional filter with the API search module
     config.memex_add_search_filter('h.nipsa.search.Filter')
